Use INT instead LONG for Min- and MaxFlightTime... in AccountingRuleFilters

+++ src/FLS.Data.WebApi/Accounting/RuleFilters/AccountingRuleFilterDetails.cs
@@ -88,11 +88,11 @@ namespace FLS.Data.WebApi.Accounting.RuleFilters
         #region invoice line rule part
         public ArticleTargetDetails ArticleTarget { get; set; }
 
-        public long? MinFlightTimeInSecondsMatchingValue { get; set; }
-        public long? MaxFlightTimeInSecondsMatchingValue { get; set; }
+        public int? MinFlightTimeInSecondsMatchingValue { get; set; }
+        public int? MaxFlightTimeInSecondsMatchingValue { get; set; }
 
-        public long? MinEngineTimeInSecondsMatchingValue { get; set; }
-        public long? MaxEngineTimeInSecondsMatchingValue { get; set; }
+        public int? MinEngineTimeInSecondsMatchingValue { get; set; }
+        public int? MaxEngineTimeInSecondsMatchingValue { get; set; }
 
         public bool IncludeThresholdText { get; set; }
